Rename Framework -> OS (in line with osu.Framework.Desktop naming).
