+++ osu.Framework/Game.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Windows.Forms;
-using osu.Framework.Framework;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Timing;
 using System.Threading;
@@ -12,6 +11,7 @@ using osu.Framework.Graphics.Textures;
 using osu.Framework.Input;
 using osu.Framework.Graphics.Shaders;
 using osu.Framework.IO.Stores;
+using osu.Framework.OS;
 using Scheduler = osu.Framework.Threading.Scheduler;
 
 namespace osu.Framework
