+++ osu.Framework/Graphics/OpenGL/Buffers/VertexBuffer.cs
@@ -2,7 +2,6 @@
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using OpenTK.Graphics.ES20;
-using osu.Framework.Framework;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
