+++ osu.Framework/osu.Framework.csproj
@@ -131,9 +131,9 @@
     <Compile Include="Audio\Track\AudioTrackBass.cs" />
     <Compile Include="Audio\Track\AudioTrackVirtual.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Game.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Framework\BasicGameHost.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Framework\BasicGameWindow.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Framework\GLControl.cs">
+    <Compile Include="OS\BasicGameHost.cs" />
+    <Compile Include="OS\BasicGameWindow.cs" />
+    <Compile Include="OS\GLControl.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Graphics\Batches\LinearBatch.cs" />
@@ -219,7 +219,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Graphics\Textures\TextureStore.cs" />
     <Compile Include="RuntimeInfo.cs" />
-    <Compile Include="Framework\BasicGameForm.cs">
+    <Compile Include="OS\BasicGameForm.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Threading\Scheduler.cs" />
